requirements.txt
tensorflow
numpy
Pillow
scipy
matplotlib
streamlit
tqdm